# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PriceCheck NG API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /best-deal:
    get:
      operationId: getBestDeal
      tags: [deals]
      summary: Get best deal options for a product (cheapest, closest, best-rated)
      parameters:
        - name: product_id
          in: query
          required: true
          schema:
            type: integer
        - name: location
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Best deal options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BestDeal"
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: Get current best deals (cheapest vendors per category)
      parameters:
        - name: category_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: location
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Current deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (optionally filter by vendor)
      parameters:
        - name: vendor_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (vendor confirms/rejects)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    BestDealOption:
      type: object
      required: [label, vendor_id, vendor_name, vendor_location, price, quantity, vendor_whatsapp]
      properties:
        label:
          type: string
        vendor_id:
          type: integer
        vendor_name:
          type: string
        vendor_location:
          type: string
        vendor_logo_url:
          type: ["string", "null"]
        vendor_verified:
          type: boolean
        vendor_whatsapp:
          type: ["string", "null"]
        vendor_rating:
          type: ["number", "null"]
        vendor_stock_available:
          type: boolean
        price:
          type: number
        quantity:
          type: string
        badge:
          type: string
    BestDeal:
      type: object
      required: [product_id, options]
      properties:
        product_id:
          type: integer
        options:
          type: array
          items:
            $ref: "#/components/schemas/BestDealOption"
    Deal:
      type: object
      required: [product_id, product_name, vendor_name, price, unit, original_avg_price]
      properties:
        product_id:
          type: integer
        product_name:
          type: string
        category_name:
          type: string
        vendor_id:
          type: integer
        vendor_name:
          type: string
        vendor_location:
          type: string
        vendor_verified:
          type: boolean
        vendor_whatsapp:
          type: ["string", "null"]
        vendor_rating:
          type: ["number", "null"]
        price:
          type: number
        unit:
          type: string
        original_avg_price:
          type: number
        discount_pct:
          type: number
        image_url:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, product_id, vendor_id, quantity, buyer_name, status, created_at]
      properties:
        id:
          type: integer
        product_id:
          type: integer
        vendor_id:
          type: integer
        product_name:
          type: ["string", "null"]
        vendor_name:
          type: ["string", "null"]
        vendor_whatsapp:
          type: ["string", "null"]
        quantity:
          type: integer
        buyer_name:
          type: string
        buyer_phone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
        price_at_order:
          type: ["number", "null"]
        created_at:
          type: string
    OrderInput:
      type: object
      required: [product_id, vendor_id, quantity, buyer_name]
      properties:
        product_id:
          type: integer
        vendor_id:
          type: integer
        quantity:
          type: integer
        buyer_name:
          type: string
        buyer_phone:
          type: string
        notes:
          type: string
        price_at_order:
          type: number
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
...