# lib/api-zod/src/generated/types/bestDeal.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PriceCheck NG API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BestDealOption } from './bestDealOption';

export interface BestDeal {
  product_id: number;
  options: BestDealOption[];
}