# lib/api-zod/src/generated/types/bestDealOption.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PriceCheck NG API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BestDealOption {
  label: string;
  vendor_id: number;
  vendor_name: string;
  vendor_location: string;
  /** @nullable */
  vendor_logo_url?: string | null;
  vendor_verified?: boolean;
  /** @nullable */
  vendor_whatsapp: string | null;
  /** @nullable */
  vendor_rating?: number | null;
  vendor_stock_available?: boolean;
  price: number;
  quantity: string;
  badge?: string;
}