# lib/api-zod/src/generated/types/deal.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PriceCheck NG API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Deal {
  product_id: number;
  product_name: string;
  category_name?: string;
  vendor_id?: number;
  vendor_name: string;
  vendor_location?: string;
  vendor_verified?: boolean;
  /** @nullable */
  vendor_whatsapp?: string | null;
  /** @nullable */
  vendor_rating?: number | null;
  price: number;
  unit: string;
  original_avg_price: number;
  discount_pct?: number;
  /** @nullable */
  image_url?: string | null;
}