# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PriceCheck NG API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  product_id: number;
  vendor_id: number;
  /** @nullable */
  product_name?: string | null;
  /** @nullable */
  vendor_name?: string | null;
  /** @nullable */
  vendor_whatsapp?: string | null;
  quantity: number;
  buyer_name: string;
  /** @nullable */
  buyer_phone?: string | null;
  /** @nullable */
  notes?: string | null;
  status: string;
  /** @nullable */
  price_at_order?: number | null;
  created_at: string;
}