# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PriceCheck NG API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  product_id: number;
  vendor_id: number;
  quantity: number;
  buyer_name: string;
  buyer_phone?: string;
  notes?: string;
  price_at_order?: number;
}